Rename cell.ts to editor.ts
